Extract EnvironmentBuilder from EnvironmentAwareCommand
Decouple generic environment bootstrapping (Docker env var
translation, system property merging, config path resolution)
into a standalone utility so non-CLI entry points can create
an Environment without inheriting from Command.
